+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.3.0
+
+* onBeforeChange now receives a (segment, src).  segment is an object.
+
+# 0.2.5
+
+* Removed instance.firstRun variable.
+
+
 # 0.2 2018-11-29T07:45
 
 * Renamed jquery.smart_images.js to smart-images.js.